Recreated the separate ItemAdapter class
