+++ app/src/main/java/king/bas/a1_comp3275/CustomAdapter.java
@@ -64,41 +64,4 @@ public class CustomAdapter extends AppCompatActivity {
 
     }
 
-    class ItemAdapter extends ArrayAdapter<String> {
-
-        public ItemAdapter(Context context, String [] items){
-            super (context, 0, items);
-        } //constructor
-
-        @Override
-        public View getView(int position, View v, ViewGroup p){
-            v = LayoutInflater.from(getContext()).inflate(R.layout.item, null);
-            String item = getItem(position);
-            String [] itemPrices = getContext().getResources().getStringArray(R.array.itemS_prices);
-            TypedArray itemImages = getContext().getResources().obtainTypedArray(R.array.items_images);
-            /*
-            final String [] itemDesc = getContext().getResources().getStringArray(R.array.items_description);
-
-            ((Button) v.findViewById(R.id.btn_deets))
-                        .setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-                            Toast.makeText(CustomAdapter.this, itemDesc[position], Toast.LENGTH_SHORT).show();
-                        }
-                    });
-*/
-            ((TextView)v.findViewById(R.id.txt_name))
-                    .setText(item);
-
-            ((TextView)v.findViewById(R.id.txt_price))
-                    .setText(itemPrices[position]);
-
-            ((ImageView)v.findViewById(R.id.img_icon))
-                    .setImageResource(itemImages.getResourceId(position, 0));
-            return v;
-
-        }// view
-
-    }
-
 }
