+++ app/src/main/res/layout/item.xml
@@ -3,7 +3,8 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="horizontal"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:weightSum="1">
 
     <ImageView
         android:layout_width="100dp"
@@ -15,7 +16,7 @@
         android:orientation="vertical"
         android:layout_width="189dp"
         android:layout_marginLeft="5dp"
-        android:layout_height="wrap_content">
+        android:layout_height="100dp">
 
         <TextView
             android:layout_width="wrap_content"
@@ -36,6 +37,7 @@
         android:layout_width="wrap_content"
         android:layout_height="63dp"
         android:text="Details"
-        android:id="@+id/btn_deets" />
+        android:id="@+id/btn_deets"
+        android:layout_weight="0.62" />
 
 </LinearLayout>
